Add more warnings

+++ main.cpp
@@ -73,7 +73,7 @@ template <typename T>
 vector<T> input_vector(size_t n)
 {
     vector<T> x(n);
-    for (int i = 0; i < n; i++)
+    for (size_t i = 0; i < n; i++)
         cin >> x[i];
 
     return x;
